LayoutTests/media/track/track-addtrack-kind-expected.txt

Tests that the addTextTrack() only accepts known 'kind' values.

++ Test invalid 'kind'.
EXPECTED (video.textTracks.length == '0') OK
TEST(video.addTextTrack('kaptions')) THROWS(DOMException.SYNTAX_ERR) OK
EXPECTED (video.textTracks.length == '0') OK

++ Test all valid 'kind' keywords.
RUN(video.addTextTrack('subtitles'))
EXPECTED (video.textTracks.length == '1') OK
RUN(video.addTextTrack('captions'))
EXPECTED (video.textTracks.length == '2') OK
RUN(video.addTextTrack('descriptions'))
EXPECTED (video.textTracks.length == '3') OK
RUN(video.addTextTrack('chapters'))
EXPECTED (video.textTracks.length == '4') OK
RUN(video.addTextTrack('metadata'))
EXPECTED (video.textTracks.length == '5') OK

END OF TEST
